Move weight creation logic
