+++ packages/api/src/weights/graph/filterNoEdge.ts
@@ -1,14 +1,14 @@
 import { Weights } from 'src/types'
 
-export default function filterNoEdge<T>(graph: Weights<T>): Weights<T> {
+export default function filterNoEdge<T>(weights: Weights<T>): Weights<T> {
   let result
-  for (const node of Object.keys(graph)) {
-    if (Object.keys(graph[node]).length > 0) {
+  for (const node of Object.keys(weights)) {
+    if (Object.keys(weights[node]).length > 0) {
       continue
     }
 
-    result = result || { ...graph }
+    result = result || { ...weights }
     delete result[node]
   }
-  return result || graph
+  return result || weights
 }
